homework 6

+++ api/src/main/java/ru/geekbrains/gainanov/market/api/OrderItemDto.java
@@ -1,17 +1,19 @@
 package ru.geekbrains.gainanov.market.api;
 
+import java.math.BigDecimal;
+
 public class OrderItemDto {
     private Long id;
     private ProductDto productDto;
     private Long orderId;
     private int quantity;
-    private int pricePerProduct;
-    private int price;
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;
 
     public OrderItemDto() {
     }
 
-    public OrderItemDto(Long id, ProductDto productDto, Long orderId, int quantity, int pricePerProduct, int price) {
+    public OrderItemDto(Long id, ProductDto productDto, Long orderId, int quantity, BigDecimal pricePerProduct, BigDecimal price) {
         this.id = id;
         this.productDto = productDto;
         this.orderId = orderId;
@@ -52,19 +54,19 @@ public class OrderItemDto {
         this.quantity = quantity;
     }
 
-    public int getPricePerProduct() {
+    public BigDecimal getPricePerProduct() {
         return pricePerProduct;
     }
 
-    public void setPricePerProduct(int pricePerProduct) {
+    public void setPricePerProduct(BigDecimal pricePerProduct) {
         this.pricePerProduct = pricePerProduct;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
